removed log_parser and added it to spotter.py
